- added util function to read current gpu load - quantized precision to quint8

+++ Fuzzer/model/StarCoder.py
@@ -84,7 +84,7 @@ class StarCoder:
     self.tokenizer = AutoTokenizer.from_pretrained(checkpoint)
     self.model = (
       GPTBigCodeForCausalLM.from_pretrained(checkpoint)
-                              .to(torch.bfloat16)
+                              .to(torch.quint8)
                               .to(self.device)
     )
     self.eos = []
